// vendor/rappid.d.ts
export namespace g {

    export interface PlainPoint {

        x: number;
        y: number;
    }

    export interface PlainRect {

        x: number;
        y: number;
        width: number;
        height: number;
    }
// ...
}
// ...
export namespace dia {

    type Event = JQuery.TriggeredEvent;

    type Point = g.PlainPoint;

    type BBox = g.PlainRect;

    type Size = Pick<BBox, 'width' | 'height'>;

    type PaddingJSON = {
        left?: number;
        top?: number;
        right?: number;
        bottom?: number;
    };

    type Padding = number | PaddingJSON;
// ...
    export namespace Graph {

        interface Options {
            [key: string]: any;
        }

        interface ConnectionOptions extends Cell.EmbeddableOptions {
            inbound?: boolean;
            outbound?: boolean;
        }

        interface ExploreOptions extends ConnectionOptions {
            breadthFirst?: boolean;
        }
    }
// ...
    export namespace Cell {

        interface GenericAttributes<T> {
            attrs?: T;
            z?: number;
            [key: string]: any;
        }

        interface Selectors {
            [selector: string]: attributes.SVGAttributes | undefined;
        }

        interface Attributes extends GenericAttributes<Selectors> {
            [key: string]: any;
        }

        interface Constructor<T extends Backbone.Model> {
            new (opt?: { id: string }): T
        }

        interface Options {
            [key: string]: any;
        }

        interface EmbeddableOptions extends Options {
            deep?: boolean;
        }

        interface DisconnectableOptions extends Options {
            disconnectLinks?: boolean;
        }

        interface TransitionOptions extends Options {
            delay?: number;
            duration?: number;
            timingFunction?: util.timing.TimingFunction;
            valueFunction?: util.interpolate.InterpolateFunction<any>;
        }
    }
// ...
}
// ...
export namespace layout {

    export namespace DirectedGraph {

        interface Edge {
            minLen?: number;
            weight?: number;
            labelpos?: 'l' | 'c' | 'r';
            labeloffset?: number;
            width?: number;
            height?: number;
        }

        interface Node {
            width?: number;
            height?: number;
        }

        interface LayoutOptions {
            dagre?: any;
            graphlib?: any;
            rankDir?: 'TB' | 'BT' | 'LR' | 'RL';
            ranker?: 'network-simplex' | 'tight-tree' | 'longest-path';
            nodeSep?: number;
            edgeSep?: number;
            rankSep?: number;
            marginX?: number;
            marginY?: number;
            resizeClusters?: boolean;
            clusterPadding?: dia.Padding;
            setPosition?: (element: dia.Element, position: dia.BBox) => void;
            setVertices?: boolean | ((link: dia.Link, vertices: dia.Point[]) => void);
            setLabels?: boolean | ((link: dia.Link, position: dia.Point, points: dia.Point[]) => void);
            debugTiming?: boolean;
            exportElement?: (element: dia.Element) => Node;
            exportLink?: (link: dia.Link) => Edge;
            // deprecated
            setLinkVertices?: boolean;
        }

        interface toGraphLibOptions {
            graphlib?: any,
            [key: string]: any
        }

        export function layout(graph: dia.Graph | dia.Cell[], opt?: LayoutOptions): g.Rect;

        export function toGraphLib(graph: dia.Graph, opt?: toGraphLibOptions): any;

        export function fromGraphLib(glGraph: any, opt?: { [key: string]: any }): dia.Graph;
    }
}
// ...
export namespace ui {
// ...
    namespace Stencil {
// ...
        export interface Group {
            label: string;
            index: number;
            closed?: boolean;
            height?: number;
            layout?: boolean | layout.GridLayout.Options | { [key: string]: any };
        }
    }
// ...
}

export namespace dia {

    export namespace Paper {

        interface PrintExportOptions {
            size?: string;
            padding?: dia.Padding;
        }

        type BeforeSerialize = (doc: SVGSVGElement, paper: dia.Paper) => void | SVGElement;

        interface RasterExportOptions {
            type?: 'image/png' | 'image/jpeg' | string;
            height?: number;
            width?: number;
            size?: string;
            backgroundColor?: string;
            quality?: number;
            padding?: dia.Padding;
            area?: dia.BBox;
            useComputedStyles?: boolean;
            stylesheet?: string;
            canvg?: any;
            beforeSerialize?: BeforeSerialize;
        }

        interface SVGExportOptions {
            preserveDimensions?: boolean;
            area?: dia.BBox;
            convertImagesToDataUris?: boolean;
            useComputedStyles?: boolean;
            stylesheet?: string;
            beforeSerialize?: BeforeSerialize;
        }
    }

    export interface Paper {

        print(opt?: Paper.PrintExportOptions): void;

        toDataURL(callback: (dataURL: string) => void, opt?: Paper.RasterExportOptions): void;

        toJPEG(callback: (dataURL: string) => void, opt?: Paper.RasterExportOptions): void;

        toPNG(callback: (dataURL: string) => void, opt?: Paper.RasterExportOptions): void;

        toSVG(callback: (svg: string) => void, opt?: Paper.SVGExportOptions): void;

        openAsSVG(opt?: Paper.SVGExportOptions): void;

        openAsPNG(opt?: Paper.RasterExportOptions): void;
    }

    namespace CommandManager {

        export interface Options {
            graph: dia.Graph;
            cmdBeforeAdd?: (eventName: string, ...eventArgs: any[]) => boolean;
            cmdNameRegex?: any; /* a regular expression */
            applyOptionsList?: string[];
            revertOptionsList?: string[];
        }
    }

    class CommandManager extends Backbone.Model {

        constructor(opt: CommandManager.Options);

        undo(opt?: { [key: string]: any }): void;

        redo(opt?: { [key: string]: any }): void;

        cancel(opt: { [key: string]: any }): void;

        reset(): void;

        hasUndo(): boolean;

        hasRedo(): boolean;

        listen(): void;

        initBatchCommand(): void;

        storeBatchCommand(): void;
    }

    namespace Validator {

        export interface Options {
            commandManager: dia.CommandManager;
            cancelInvalid?: boolean;
        }
    }

    class Validator extends Backbone.Model {

        constructor(opt: Validator.Options);

        validate(actions: string, ...callback: Array<(err: Error, command: any, next: any) => boolean>): Validator;
    }
}
// ...
export namespace layout {

    namespace ForceDirected {
        export interface Options {
            graph: dia.Graph | dia.Cell[];
            x?: number;
            y?: number;
            width: number;
            height: number;
            gravityCenter: dia.Point;
            charge?: number;
            linkDistance?: number;
            linkStrength?: number;
        }
    }

    class ForceDirected extends Backbone.Model {

        constructor(opt: ForceDirected.Options);

        start(): void;

        step(): void;
    }

    namespace GridLayout {

        export interface Options {
            centre?: boolean;
            dx?: number;
            dy?: number;
            resizeToFit?: boolean;
            marginX?: number;
            marginY?: number;
            columns?: number;
            columnWidth?: 'compact' | 'auto' | number | string;
            rowHeight?: 'compact' | 'auto' | number | string;
            deep?: boolean;
            parentRelative?: boolean;
        }

        export function layout(graphOrCells: dia.Graph | Array<dia.Cell>, opt?: Options): void;
    }

    namespace TreeLayout {

        interface AttributeNames {
            'siblingRank'?: string;
            'direction'?: string;
            'margin'?: string;
            'offset'?: string;
            'prevSiblingGap'?: string;
            'nextSiblingGap'?: string;
            'firstChildGap'?: string;
        }

        type Direction = 'L' | 'R' | 'T' | 'B' | 'BR' | 'BL' | 'TR' | 'TL';

        type DirectionRule = (rule: [Direction, Direction]) => (direction: Direction) => Direction;

        export interface DirectionRules {
            rotate: DirectionRule;
            flip: DirectionRule;
            straighten: DirectionRule;
        }

        export interface Options {
            graph: dia.Graph;
            gap?: number;
            parentGap?: number;
            siblingGap?: number;
            firstChildGap?: number;
            direction?: Direction;
            directionRule?: DirectionRule;
            updatePosition?: null | ((element: dia.Element, position: dia.Point, opt?: { [key: string]: any }) => void);
            updateVertices?: null | ((link: dia.Link, vertices: Array<dia.Point>, opt?: { [key: string]: any }) => void);
            updateAttributes?: null | ((layoutArea: any, root: dia.Element, rootLink: dia.Link, opt: { [key: string]: any }) => void);
            filter?: null | ((children: dia.Element[], parent: dia.Element | null, opt: { [key: string]: any }) => dia.Element[]);
            attributeNames?: AttributeNames;
        }
    }

    class TreeLayout extends Backbone.Model {

        constructor(opt: TreeLayout.Options);

        layout(opt?: { [key: string]: any }): this;

        layoutTree(root: dia.Element, opt?: { [key: string]: any }): this;

        getLayoutBBox(): g.Rect | null;

        static directionRules: TreeLayout.DirectionRules
    }
}